% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot Helper Functions.R
\name{add_label}
\alias{add_label}
\title{Adds label to plot}
\usage{
add_label(xfrac, yfrac, label, pos = 4, ...)
}
\arguments{
\item{xfrac}{The fraction over from the left side.}

\item{yfrac}{The fraction down from the top.}

\item{label}{The text to label with.}

\item{pos}{Position to pass to text()}

\item{...}{Anything extra to pass to text(), e.g. cex, col.}
}
\description{
Adds label to plot
}
