% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{avg_widths}
\alias{avg_widths}
\title{Calculates average of top and bottom channel width}
\usage{
avg_widths(path = "", plot = FALSE)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{plot}{Should channel width be plotted (defaults to `FALSE`)}
}
\value{
A list of average channel widths by reach
}
\description{
Calculates average of top and bottom channel width
}
