% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grain_Size_Calculators.R
\name{calc_Dx}
\alias{calc_Dx}
\title{Calculates grain size statistic from given distribution}
\usage{
calc_Dx(ps, Ds, x)
}
\arguments{
\item{ps}{Grain size fractions}

\item{Ds}{Grain sizes (mm)}

\item{x}{Percetile to find (e.g. 0.5 for $D_{50}$)}
}
\value{
Size of $D_x$
}
\description{
Calculates $D_{x}$ from a grain size distribtion, where `x` is the fraction
of the grain size distribution finer than the calculated grain size (e.g. `x
= 0.5` for $D_{50}$).
}
