% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grain_Size_Calculators.R
\name{calc_omega_c}
\alias{calc_omega_c}
\title{Calculates critical stream power for the initial grain size distribution, by reach}
\usage{
calc_omega_c(path, return_vals = FALSE)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{return_vals}{Logical. If TRUE, returns calculated critical stream powers by grain size and reach}
}
\value{
Critical stream power by grain size
}
\description{
Calculates critical stream power for the initial grain size distribution, by reach
}
