% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grain_Size_Calculators.R
\name{calc_sigmag}
\alias{calc_sigmag}
\title{Calculates geometric standard deviation of grain size distribtion}
\usage{
calc_sigmag(ps, Ds)
}
\arguments{
\item{ps}{Grain size fractions}

\item{Ds}{Grain sizes (mm)}
}
\value{
Geometric standard deviation
}
\description{
Calculates geometric standard deviation of grain size distribtion
}
