% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot Helper Functions.R
\name{cRamp}
\alias{cRamp}
\title{Creates a color ramp from a set of values}
\usage{
cRamp(x, palette, alpha = 1)
}
\arguments{
\item{x}{Series of values to create color ramp for}

\item{palette}{Palette of colors to use (either `viridis` or a palette from
`RColorBrewer`)}

\item{alpha}{Transparency factor (defaults to `1`)}
}
\value{
Set of colors corresponding to each supplied value
}
\description{
Creates a color ramp from a set of values
}
