% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot Helper Functions.R
\name{cRamp_legend}
\alias{cRamp_legend}
\title{Creates a color ramp of specified length}
\usage{
cRamp_legend(x, palette, alpha = 1)
}
\arguments{
\item{x}{Length of color palette}

\item{palette}{Palette of colors to use (either `viridis` or a palette from
`RColorBrewer`)}

\item{alpha}{Transparency factor (defaults to `1`)}
}
\value{
Set of colors of length `x`
}
\description{
Creates a color ramp of specified length
}
