% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{D50_plot}
\alias{D50_plot}
\title{Plots changes in bed median grain size over time for all cross sections}
\usage{
D50_plot(path = "")
}
\arguments{
\item{path}{Path to folder with model outputs}
}
\description{
Plots changes in bed median grain size over time for all cross sections
}
