% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{data_by_XS}
\alias{data_by_XS}
\title{Transforms model outputs into data by each cross section}
\usage{
data_by_XS(data)
}
\arguments{
\item{data}{A matrix of model output data}
}
\value{
A list of data by cross section
}
\description{
Transforms model outputs into data by each cross section
}
