% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{dz_plot}
\alias{dz_plot}
\title{Plots changes in bed elevation for each cross section in the network}
\usage{
dz_plot(print = FALSE, gif = FALSE, max_plots = 10, path = "",
  custom_sgn = NULL, title = NULL)
}
\arguments{
\item{print}{Should the plot be printed (defaults to `FALSE`)}

\item{gif}{Should a gif be created (defaults to `FALSE`)}

\item{max_plots}{Maximum number of plots in gif}

\item{path}{Path to folder with model outputs}

\item{custom_sgn}{Specifies the direction each reach should be plotted (`-1`
is left, `1` is right)}

\item{title}{Title to be printed on plot}
}
\description{
Plots changes in bed elevation for each cross section in the network
}
