% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grain_Size_Calculators.R
\name{gsd_maker}
\alias{gsd_maker}
\title{Creates a grain size distribution}
\usage{
gsd_maker(D50, sp, ds, plot = TRUE)
}
\arguments{
\item{D50}{Median grain size (mm).}

\item{sp}{Spread of distribution (default is 1).}

\item{ds}{Vector of grain sizes to map the gsd to (mm).}

\item{plot}{Should the GSD be plotted (default is `TRUE``).}
}
\value{
The size fraction for each grain size.
}
\description{
Creates a grain size distribution, given a set of grain sizes, $D_{50}$, and
sp.
}
