% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{knickpoint_plot}
\alias{knickpoint_plot}
\title{Plots knickpoint location over time}
\usage{
knickpoint_plot(path = "")
}
\arguments{
\item{path}{Path to folder with model outputs}
}
\description{
Plots knickpoint location over time
}
