% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{make_network}
\alias{make_network}
\title{Creates a series of points visualizing the channel network}
\usage{
make_network(n_nodes, n_xs, link, dx, custom_sgn)
}
\arguments{
\item{n_nodes}{The number of channel reaches or nodes}

\item{n_xs}{Number of cross sections per reach}

\item{link}{Matrix specifying reach layout}

\item{dx}{Cross section spacing}

\item{custom_sgn}{Specifies the direction each reach should be plotted (`-1`
is left, `1` is right)}
}
\value{
List of x and y coordinates of channel network
}
\description{
Creates a series of points visualizing the channel network
}
