% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{network_XS_plot}
\alias{network_XS_plot}
\title{Plots the channel network showing changes in bed elevation, width, and width-depth ratio over time}
\usage{
network_XS_plot(path = "", XS = NULL, pos = c("right", "right"),
  years = c(1, 3, 5, 10, 20), print = FALSE)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{XS}{Number of cross sections to label}

\item{pos}{Position of labels for cross sections (`right` or `left`)}

\item{years}{A numeric vectors of years of the simulation to plot}

\item{print}{Should the plot be printed to a file (defaults to `FALSE`)}
}
\description{
Plots the channel network showing changes in bed elevation, width, and width-depth ratio over time
}
