% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot Helper Functions.R
\name{plot_colors}
\alias{plot_colors}
\title{A series of nice, color-blind friendly, colors for plotting}
\usage{
plot_colors(alpha = 1, plot = FALSE)
}
\arguments{
\item{alpha}{Transparency factor (defaults to `1`)}

\item{plot}{Should the colors be plotted (defaults to `FALSE`)}
}
\value{
A set of eight named colors
}
\description{
A series of nice, color-blind friendly, colors for plotting
}
