% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{plot_omega}
\alias{plot_omega}
\title{Plots specific stream power}
\usage{
plot_omega(path = "", type = 1)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{type}{`type = 1` plots stream power over time for each reach
separately; `type = 2` plots stream power longitudinally by reach}
}
\description{
Plots specific stream power
}
