% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{Rc_lines}
\alias{Rc_lines}
\title{Plots changes in bend radius of curvature over time for all cross sections}
\usage{
Rc_lines(path = "")
}
\arguments{
\item{path}{Path to folder with model outputs}
}
\description{
Plots changes in bend radius of curvature over time for all cross sections
}
