% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{reach_loads}
\alias{reach_loads}
\title{Plots mass sediment loading rates by reach, with uncertainty}
\usage{
reach_loads(path = "", custom_sgn = NULL, MC_path = NULL, n_MC = 0,
  units = "ton", prob = c(0.05, 0.95), print = FALSE)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{custom_sgn}{Specifies the direction each reach should be plotted (`-1`
is left, `1` is right)}

\item{MC_path}{Path to "MC Outputs" folder (only if different than `path`)}

\item{n_MC}{Number of Monte Carlo simulations}

\item{units}{Character specifying units to be used in plot ("kg", "ton", or "1000 ton")}

\item{prob}{Numeric vector of percentiles of Monte Carlo results to plot in
addition to the median (defaults to 0.05 and 0.95)}

\item{print}{Should the plot be printed to a file (defaults to `FALSE`)}
}
\description{
Plots mass sediment loading rates by reach, with uncertainty
}
