% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{sed_lines}
\alias{sed_lines}
\title{Plots sediment inflow and outflow over time}
\usage{
sed_lines(path = "")
}
\arguments{
\item{path}{Path to folder with model outputs}
}
\description{
Plots sediment inflow and outflow over time
}
