% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{sinuosity_plot}
\alias{sinuosity_plot}
\title{Plots changes in channel sinuosity over time by reach}
\usage{
sinuosity_plot(path = "")
}
\arguments{
\item{path}{Path to folder with model outputs}
}
\description{
Plots changes in channel sinuosity over time by reach
}
