% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{slope_lines}
\alias{slope_lines}
\title{Plots changes in channel slope over time for all cross sections}
\usage{
slope_lines(path = "")
}
\arguments{
\item{path}{Path to folder with model outputs}
}
\description{
Plots changes in channel slope over time for all cross sections
}
