% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{width_depth}
\alias{width_depth}
\title{Calculates channel width-depth ratio}
\usage{
width_depth(path = "", plot = FALSE)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{plot}{Should the data be plotted (defaults to `FALSE`)}
}
\value{
A dataframe with width-depth ratio by reach
}
\description{
Calculates channel width-depth ratio
}
