% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{width_MC_plot}
\alias{width_MC_plot}
\title{Plots network showing changes in channel width, with uncertainty}
\usage{
width_MC_plot(print = FALSE, n_MC, path = "", MC_path = NULL,
  custom_sgn = NULL, prob = c(0.05, 0.95))
}
\arguments{
\item{print}{Should the plot be printed to a file (defaults to `FALSE`)}

\item{n_MC}{Number of Monte Carlo simulations}

\item{path}{Path to folder with model outputs}

\item{MC_path}{Path to "MC Outputs" folder (only if different than `path`)}

\item{custom_sgn}{Specifies the direction each reach should be plotted (`-1`
is left, `1` is right)}

\item{prob}{Numeric vector of percentiles of Monte Carlo results to plot in
addition to the median (defaults to 0.05 and 0.95)}
}
\description{
Plots network showing changes in channel width, with uncertainty
}
