% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{width_plot}
\alias{width_plot}
\title{Plots changes in channel width for all cross sections in the network}
\usage{
width_plot(print = FALSE, gif = FALSE, max_plots = 10, path = "",
  custom_sgn = NULL, title = NULL)
}
\arguments{
\item{gif}{Should a gif be created (defaults to `FALSE`)}

\item{max_plots}{Maximum number of plots in gif}

\item{path}{Path to folder with model outputs}

\item{custom_sgn}{Specifies the direction each reach should be plotted (`-1`
is left, `1` is right)}

\item{title}{Title to be printed on plot}
}
\description{
#' @param print Should the plot be printed (defaults to `FALSE`)
}
