% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{XS_areas}
\alias{XS_areas}
\title{Calculates a mass balance of modeled channel change}
\usage{
XS_areas(path = "")
}
\arguments{
\item{path}{Path to folder with model outputs}
}
\value{
Prints results of the mass balance and a plot of volume changes of
  channel cross sections.

  \tabular{ll}{ Volume sum \tab Sum of total channel volume change ((-)
  indicates net erosion, (+) indicates net aggradation)\cr Bed vol out \tab
  Total volume of bed material load explorted from the watershed\cr Bed vol
  in \tab Total volume of bed material load imported to watershed\cr Bank
  tank \tab Volume of failed bank material in the bank "tank" (this is
  material that couldn't be deposited at the bank toe but is stored in a
  virtual "tank" to conserve mass.)\cr Bank washload \tab Volume of eroded
  bank washload\cr Bed washload (cohesive) \tab Volume of eroded washload
  from cohesive bed erosion\cr Knickpoint washload \tab Volume of eroded
  washload from knickpoint erosion\cr Knickpoint correction \tab A volume
  correction for when a knickpoint is initially located between two cross
  sections\cr Diff \tab Calculated volume difference between calculated
  change in sediment inflow and outflow and total channel change\cr Percent
  Diff \tab Calculated volume difference as a percentage of Volume sum\cr }
}
\description{
Calculates changes in cross section area and compares that to sediment inputs
and outputs to determine if mass was conserved during the simulation. Note
the mass balance is not accurate if meandering was simulated.
}
