% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{XS_plots}
\alias{XS_plots}
\title{Plots initial and final cross section geometry for the most upstream cross
section in each reach}
\usage{
XS_plots(path = "")
}
\arguments{
\item{path}{Path to folder with model outputs}
}
\description{
Plots initial and final cross section geometry for the most upstream cross
section in each reach
}
