% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{XS_plots2}
\alias{XS_plots2}
\title{Plots initial and final cross section geometry for any specified cross sections}
\usage{
XS_plots2(path = "", XS = 1)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{XS}{A numeric vector of the cross sections to be plotted}
}
\description{
Plots initial and final cross section geometry for any specified cross sections
}
