% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{XS_plots3}
\alias{XS_plots3}
\title{Plots cross section geometry over time for any specified cross section}
\usage{
XS_plots3(path = "", XS = 1, n_plots = 0, ts = 0.2, print = FALSE)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{XS}{The number of the cross section to be plotted}

\item{n_plots}{The number of cross sections to plot}

\item{ts}{The time step (in seconds) between plottings}

\item{print}{Should the plot be printed to a file (defaults to `FALSE`)}
}
\description{
Plots cross section geometry over time for any specified cross section
}
