% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{pollutant_loading}
\alias{pollutant_loading}
\title{Plots cumulative or daily pollutant loading}
\usage{
pollutant_loading(path = "", type = 1)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{type}{`type = 1` plots cumulative loads, `type = 2` plots daily loads}
}
\description{
Plots cumulative or daily pollutant loading
}
