% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization Functions.R
\name{width_lines}
\alias{width_lines}
\title{Plots changes in channel width over time for all cross sections}
\usage{
width_lines(path = "", print = FALSE)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{print}{Should the plot be printed to a file (defaults to `FALSE`)}
}
\description{
Plots changes in channel width over time for all cross sections
}
