% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_Functions.R
\name{XS_plots2}
\alias{XS_plots2}
\title{Plots initial and final cross section geometry for any specified cross sections}
\usage{
XS_plots2(path = "", reach = 1, XS = 1)
}
\arguments{
\item{path}{Path to folder with model outputs}

\item{reach}{A numeric vector of reaches for which to plot cross sections}

\item{XS}{A numeric vector of the cross sections within each reach to be
plotted (one `XS` per `reach`)}
}
\description{
Plots initial and final cross section geometry for any specified cross sections
}
